Refactor to streamline service imports and classes
Updated multiple service import paths to use consistent naming conventions. Simplified the AuctionService by introducing a BaseService parent class to handle common CRUD operations. Adjusted relationship attributes for clarity in models.
