+++ python_auctioneer/menus/auction_menu.py
@@ -1,9 +1,10 @@
 """Manage auctions menu."""
 
 import datetime
-from database import SessionLocal
-from python_auctioneer.services.auction import create_auction_service, get_auctions_service, update_auction_service, \
-    get_auction_by_id, delete_auction_service
+from database import SessionLocal, get_database
+from python_auctioneer.models import Auction
+from python_auctioneer.services.auction import AuctionService
+
 from tabulate import tabulate
 
 MENU = """Auction Management
@@ -20,16 +21,16 @@ def auction_menu():
     print(MENU)
     choice = input(">> ")
     while choice != "5":
-        if choice == "1":
-            create_auction(database)
-        elif choice == "2":
-            display_auctions(get_auctions_service(database))
-        elif choice == "3":
-            update_auction(database)
-            pass
-        elif choice == "4":
-            delete_auction(database)
-            pass
+        if choice in ["1", "2", "3", "4"]:
+            with get_database() as database:
+                if choice == "1":
+                    create_auction(database)
+                elif choice == "2":
+                    display_auctions(database)
+                elif choice == "3":
+                    update_auction(database)
+                else:
+                    delete_auction(database)
         else:
             print("Invalid choice.")
         print(MENU)
@@ -51,15 +52,17 @@ def create_auction(database):
             "auction_open_time": auction_open_time,
             "auction_close_time": auction_close_time,
         }
-
-        create_auction_service(database, auction_data)
+        auction_service = AuctionService()
+        auction_service.create(database, auction_data)
     except ValueError as e:
         print(f"Error: {e}")
         print("Invalid date format. Please use DD/MM/YYYY.")
 
 
-def display_auctions(auctions):
+def display_auctions(database):
     """Print a formatted table of auctions."""
+    auction_service = AuctionService()
+    auctions = auction_service.get_all(database)
     if not auctions:
         print("No auctions found.")
     else:
@@ -78,7 +81,8 @@ def display_auctions(auctions):
 
 def update_auction(database):
     auction_id = input("Enter the auction ID to update: ")
-    auction = get_auction_by_id(database, auction_id)
+    auction_service = AuctionService()
+    auction = auction_service.get_by_id(database, auction_id)
 
     if not auction:
         print("Auction not found.")
@@ -120,7 +124,8 @@ def delete_auction(database):
     auction_id = input("Enter the auction ID to delete: ")
     confirmation = input(f"Are you sure you want to delete auction ID {auction_id}? (y/n): ")
     if confirmation.lower() == 'y':
-        success = delete_auction_service(database, auction_id)
+        auction_service = AuctionService()
+        success = auction_service.delete(database, auction_id)
         if success:
             print("Auction deleted successfully.")
         else:
