+++ python_auctioneer/models/finish.py
@@ -1,6 +1,5 @@
 from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import relationship
-
 from python_auctioneer.models import Base
