+++ python_auctioneer/services/shipping.py
@@ -1,9 +1,9 @@
 from sqlalchemy.exc import IntegrityError
-from python_auctioneer.models.shipping_method import ShippingMethod
+from python_auctioneer.models.shipping import Shipping
 
 def add_shipping_method_service(database, shipping_method_data):
     try:
-        database_shipping_method = ShippingMethod(**shipping_method_data)
+        database_shipping_method = Shipping(**shipping_method_data)
         database.add(database_shipping_method)
         database.commit()
         database.refresh(database_shipping_method)
